test fixes

+++ tests/models/charge_measurement/test_nghxrg.py
@@ -36,7 +36,7 @@ def cmos_10x10() -> CMOS:
 
     # Generate fake pixels
     rs = np.random.RandomState(12345)
-    detector.pixel.array = rs.normal(loc=10000, scale=500, size=(10, 10)).astype(int)
+    detector.pixel.array = rs.normal(loc=10000, scale=500, size=(10, 10))
 
     return detector
 
@@ -59,7 +59,7 @@ def cmos_10x15() -> CMOS:
 
     # Generate fake pixels
     rs = np.random.RandomState(12345)
-    detector.pixel.array = rs.normal(loc=10000, scale=500, size=(10, 15)).astype(int)
+    detector.pixel.array = rs.normal(loc=10000, scale=500, size=(10, 15))
 
     return detector
